# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MealFlow API
  description: India-focused family meal planning app API
  version: 0.9.0
  license:
    name: MIT
servers:
  - url: http://localhost:3458
    description: Local development
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: mf_sid
  schemas:
...
    Ingredient:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        calories:
          type: number
        protein:
          type: number
        carbs:
          type: number
        fat:
          type: number
        fiber:
          type: number
        unit:
          type: string
        created_at:
          type: string
          format: date-time
...
    Household:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        created_by:
          type: integer
        members:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              email:
                type: string
              display_name:
                type: string
        person_count:
          type: integer
...
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
security:
  - cookieAuth: []
paths:
...
  /api/ingredients:
    get:
      tags: [ingredients]
      summary: List ingredients (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: sort
          in: query
          schema:
            type: string
            enum: [name, category, calories]
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: Paginated ingredient list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Ingredient'
    post:
      tags: [ingredients]
      summary: Create an ingredient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, category]
              properties:
                name:
                  type: string
                category:
                  type: string
                calories:
                  type: number
                protein:
                  type: number
                carbs:
                  type: number
                fat:
                  type: number
                fiber:
                  type: number
                unit:
                  type: string
      responses:
        '201':
          description: Ingredient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
  /api/ingredients/{id}:
    get:
      tags: [ingredients]
      summary: Get ingredient by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Ingredient detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
    put:
      tags: [ingredients]
      summary: Update ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                category:
                  type: string
                calories:
                  type: number
      responses:
        '200':
          description: Updated ingredient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ingredient'
    delete:
      tags: [ingredients]
      summary: Delete ingredient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
...
  /api/households:
    post:
      tags: [households]
      summary: Create a household
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '201':
          description: Household created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
  /api/households/current:
    get:
      tags: [households]
      summary: Get current user's household with members and person count
      responses:
        '200':
          description: Household detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
    put:
      tags: [households]
      summary: Update household name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
    delete:
      tags: [households]
      summary: Delete household (admin only)
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
...
  /api/households/join:
    post:
      tags: [households]
      summary: Join household by invite code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        '200':
          description: Joined
          content:
            application/json:
              schema:
                type: object
                properties:
                  household_id:
                    type: integer
...